+++ package.json
@@ -1,6 +1,6 @@
 {
   "name": "pillarx",
-  "version": "0.12.0",
+  "version": "0.12.1",
   "private": true,
   "scripts": {
     "start": "react-scripts start",
