+++ src/components/BottomMenu/index.tsx
@@ -125,7 +125,6 @@ const MenuItem = styled.div`
 `;
 
 const Wrapper = styled.div`
-  // max-width: 800px; TODO: inspect and remove/adjust
   margin: 0 auto;
   display: flex;
   align-items: center;
